refactor: consolidate economy preview tests to three files (#3876)
Merge assignments and consumption coverage into economy_stockpile_preview_test.dart
with shared pending-work scenario helpers split across support modules to stay
under the 400-line domain test file cap.

Refs #3876
